+++ src/ballistica/base/networking/network_reader.cc
@@ -133,7 +133,6 @@ void NetworkReader::DoSelect(bool* can_read_4, bool* can_read_6) {
       // Try to get a clean error instead of a crash if we exceed our
       // open file descriptor limit (except on windows where FD_SETSIZE
       // is apparently a dummy value).
-      CheckFDThreshold(sd4_);
       if (sd4_ < 0 || sd4_ >= FD_SETSIZE) {
         FatalError("Socket/File Descriptor Overflow (sd4="
                    + std::to_string(sd4_) + ", FD_SETSIZE="
@@ -148,7 +147,6 @@ void NetworkReader::DoSelect(bool* can_read_4, bool* can_read_6) {
       // Try to get a clean error instead of a crash if we exceed our
       // open file descriptor limit (except on windows where FD_SETSIZE
       // is apparently a dummy value).
-      CheckFDThreshold(sd6_);
       if (sd6_ < 0 || sd6_ >= FD_SETSIZE) {
         FatalError("Socket/File Descriptor Overflow (sd6="
                    + std::to_string(sd6_) + ", FD_SETSIZE="
@@ -172,27 +170,6 @@ void NetworkReader::DoSelect(bool* can_read_4, bool* can_read_6) {
     *can_read_4 = sd4_ != -1 && FD_ISSET(sd4_, &readset);
     *can_read_6 = sd6_ != -1 && FD_ISSET(sd6_, &readset);
   }
-}
-
-void NetworkReader::CheckFDThreshold(int val) {
-  if (passed_fd_threshold_) {
-    return;
-  }
-
-  // Let's trigger when we pass 2/3 of the FD limit.
-  if (val < FD_SETSIZE * 2 / 3) {
-    return;
-  }
-
-  // If we pass the threshold, do a one-time dump of info
-  // to try and debug it.
-  passed_fd_threshold_ = true;
-  g_base->logic->event_loop()->PushCall([] {
-    assert(g_base->InLogicThread());
-    g_base->python->objs()
-        .Get(BasePython::ObjID::kOnTooManyFileDescriptorsCall)
-        .Call();
-  });
 }
 
 auto NetworkReader::RunThread() -> int {