+++ src/ballistica/core/platform/core_platform.cc
@@ -489,6 +489,8 @@ auto CorePlatform::IsRunningOnDesktop() -> bool {
 }
 
 void CorePlatform::SleepMillisecs(millisecs_t ms) {
+  // If we're holding the Python GIL, release it while we sleep.
+  Python::ScopedInterpreterLockRelease release;
   std::this_thread::sleep_for(std::chrono::milliseconds(ms));
 }
 
@@ -849,10 +851,6 @@ void CorePlatform::V1LoginDidChange() {
 
 void CorePlatform::SignOutV1() {
   Log(LogLevel::kError, "SignOutV1() unimplemented");
-}
-
-void CorePlatform::AndroidShowWifiSettings() {
-  Log(LogLevel::kError, "AndroidShowWifiSettings() unimplemented");
 }
 
 void CorePlatform::SetHardwareCursorVisible(bool visible) {